Fixed the previous fuckup when I accidentally put tests into main.py
